+++ martial-arts-studio-manager-client/src/app/pages/items/page.tsx
@@ -1,45 +1,45 @@
 'use client';
 
-import { useState } from 'react';
+import React, { useState } from 'react';
 import { Container, Button, Modal } from 'react-bootstrap';
-import EquipmentList from '../components/EquipmentList';
-import EquipmentForm from '../components/EquipmentForm';
-import { Equipment } from '../types/Equipment';
+import { ItemList } from '../../components/ItemList';
+import { ItemForm } from '../../components/ItemForm';
+import { Item } from '../../types/Item';
 
-export default function EquipmentPage() {
+export default function ItemsPage() {
     const [showForm, setShowForm] = useState(false);
-    const [selectedEquipment, setSelectedEquipment] = useState<Equipment | undefined>();
+    const [selectedItem, setSelectedItem] = useState<Item | undefined>();
 
     const handleClose = () => {
         setShowForm(false);
-        setSelectedEquipment(undefined);
+        setSelectedItem(undefined);
     };
 
-    const handleEdit = (equipment: Equipment) => {
-        setSelectedEquipment(equipment);
+    const handleEdit = (item: Item) => {
+        setSelectedItem(item);
         setShowForm(true);
     };
 
     return (
-        <Container fluid>
+        <Container className="py-4">
             <div className="d-flex justify-content-between align-items-center mb-4">
-                <h1>Equipment Management</h1>
+                <h1>Items</h1>
                 <Button variant="primary" onClick={() => setShowForm(true)}>
-                    Add New Equipment
+                    Add New Item
                 </Button>
             </div>
 
-            <EquipmentList onEdit={handleEdit} />
+            <ItemList onEdit={handleEdit} />
 
             <Modal show={showForm} onHide={handleClose} size="lg">
                 <Modal.Header closeButton>
                     <Modal.Title>
-                        {selectedEquipment ? 'Edit Equipment' : 'Add New Equipment'}
+                        {selectedItem ? 'Edit Item' : 'Add New Item'}
                     </Modal.Title>
                 </Modal.Header>
                 <Modal.Body>
-                    <EquipmentForm 
-                        equipment={selectedEquipment} 
+                    <ItemForm
+                        item={selectedItem}
                         onSuccess={handleClose}
                     />
                 </Modal.Body>
