+++ martial-arts-studio-manager-client/src/app/pages/login/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React from 'react';
-import { Login } from '../components/auth/Login';
+import { Login } from '../../components/auth/Login';
 
 export default function LoginPage() {
     return <Login />;
