+++ martial-arts-studio-manager-client/src/app/pages/register/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React from 'react';
-import { Register } from '../components/auth/Register';
+import { Register } from '../../components/auth/Register';
 
 export default function RegisterPage() {
     return <Register />;
