+++ martial-arts-studio-manager-client/src/app/pages/students/page.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import React from 'react';
-import StudentList from '../components/StudentList';
+import StudentList from '../../components/StudentList';
 
 export default function StudentsPage() {
     return <StudentList />;
