+++ martial-arts-studio-manager-client/tsconfig.json
@@ -19,7 +19,7 @@
       }
     ],
     "paths": {
-      "src/*": ["./src/*"]
+      "@/*": ["./src/*"]
     }
   },
   "include": ["next-env.d.ts", "**/*.ts", "**/*.tsx", ".next/types/**/*.ts"],
